Extract core/ domain types (step 2 of layered refactor)
Move Screen/PairingFocus/HomeFocus, MenuEvent, and the plain data
structs (ConnectTarget, KnownHost, Settings + its enums, GameEntry,
Artwork) into src/core, with pub-use re-exports from app/mod.rs,
ui/input.rs, services/store.rs, and services/library.rs so existing
call sites keep working unchanged. No behavior change.
