+++ src/core/event.rs
@@ -1,0 +1,12 @@
+/// Menu event (debounced from raw SDL2 input: keyboard arrows, gamepad d-pad).
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum MenuEvent {
+    Up,
+    Down,
+    Left,
+    Right,
+    Confirm,
+    Back,
+    /// "Forget host" (separate from Back/Confirm to prevent accident).
+    Secondary,
+}
