+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod event;
+pub mod model;
+pub mod screen;
